+++ lib/ace/edit_session.js
@@ -969,8 +969,7 @@ var EditSession = function(text, mode) {
         try {
             this.$worker = this.$mode.createWorker(this);
         } catch (e) {
-            if (typeof console == "object" && console.error)
-                console.log("Could not load worker", e);
+            config.warn("Could not load worker", e);
             this.$worker = null;
         }
     };
