+++ lib/ace/tokenizer.js
@@ -31,6 +31,7 @@
 define(function(require, exports, module) {
 "use strict";
 
+var config = require("./config");
 // tokenizing lines longer than this makes editor very slow
 var MAX_TOKEN_COUNT = 2000;
 /**
@@ -353,13 +354,8 @@ var Tokenizer = function(rules) {
         };
     };
     
-    this.reportError = function(msg, data) {
-        var e = new Error(msg);
-        e.data = data;
-        if (typeof console == "object" && console.error)
-            console.error(e);
-        setTimeout(function() { throw e; });
-    };
+    this.reportError = config.reportError;
+    
 }).call(Tokenizer.prototype);
 
 exports.Tokenizer = Tokenizer;
